// UI/regood/src/app/model/sign-up-model.ts
export class SignUp {
    username: string;
    password: string;
    email: string;
    full_name: string;
    phone_number: string;
    // dev_key: string = "3fc9b689459d738f8c88a3a48aa9e33542016b7a4052e001aaa536fca74813cb";
    // website_name: string ="reGood" ;
    city: string;
    state: string;
    zip: string;
    
    signUpModel(){
        
    }
}